% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{c24}
\alias{c24}
\title{c24}
\format{
An object of class \code{character} of length 24.
}
\usage{
c24
}
\description{
A color palatte containing 24 different colors for visualization
}
\author{
Tianxiao Zhao \email{Tianxiao.Zhao@nyulangone.org}
}
\keyword{color}
\keyword{palatte}
