% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellAlignment.R
\name{CalcCos}
\alias{CalcCos}
\title{Calculate the cosine similarity between all scRNA-seq cells and one ST spots}
\usage{
CalcCos(sc.data, ST.data.column)
}
\arguments{
\item{sc.data}{A data.frame indicating the scRNA-seq data; each column is a cell and each row is a gene}

\item{ST.data.column}{A vector containing the gene expression of a given spot}
}
\value{
A vector of cosine similarity between cells and one spot
}
\description{
A brief function to calculate the cosine similarity between all cells compared to one spot in ST data
}
