% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kriging.R
\name{createGridDataframe}
\alias{createGridDataframe}
\title{Create finer grid across the space}
\usage{
createGridDataframe(spatial.data, predict.grid.size = 20)
}
\arguments{
\item{spatial.data}{Integrated ST data from integrateCoordinate.}

\item{predict.grid.size}{The size of the finer grid; a numeric value used in createGridDataframe}
}
\value{
A data.frame containing "X" and "Y" as column names; each row is one tile in the grid networks.
}
\description{
Third sub-function of SpatialKriging. Create a finer grid for the Kriging interpolation. In general, it generates a network with a n-micron by n-micron square grid. The Kriging model can then do the interpolation based on this finer grid.
}
