% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellAlignment.R
\name{DecomposeSpot}
\alias{DecomposeSpot}
\title{Decompose the ST spot into several cells}
\usage{
DecomposeSpot(sc.data, ST.data, spot.number, thres = 0.95, iter.max = 50)
}
\arguments{
\item{sc.data}{A data.frame indicating the scRNA-seq data; each column is a cell and each row is a gene}

\item{ST.data}{A data.frame indicating the ST data; each column is a spot and each row is a gene}

\item{spot.number}{The number of spot to choose; should be an integer indicating the column number in ST data}

\item{thres}{Threshold for the sum correlation, used as a threshold for early-stop; a numeric value form -1 to 1}

\item{iter.max}{Number of max iterations, used as a threshold for early-stop; an integer larger than 1}
}
\value{
A list containing the results for one spot: which cells are chosen, and what is the predicted expression profile
}
\description{
Based on the correlation, perform the forward selecting to decompose the spot by cells from scRNA-seq data;
For one spot, do the iteration until the general correlation is larger than 0.8 or doesn't increase
}
