% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellAlignment.R
\name{DecomposeST}
\alias{DecomposeST}
\title{Decompose the ST data by estimating the composition of cells in spots}
\usage{
DecomposeST(sc.data, ST.data, thres = 0.95, iter.max = 50)
}
\arguments{
\item{sc.data}{A data.frame indicating the scRNA-seq data; each column is a cell and each row is a gene}

\item{ST.data}{A data.frame indicating the ST data; each column is a spot and each row is a gene}

\item{thres}{Threshold for the sum correlation, used as a threshold for early-stop; a numeric value form -1 to 1}

\item{iter.max}{Number of max iterations, used as a threshold for early-stop; an integer larger than 1}
}
\value{
A list containing the results for each spot: which cells are chosen, and what is the predicted expression profile
}
\description{
Perform the decomposition by calculating the correlations between scRNA-seq cells and spots' expression profile, forward-selecting the subset of cells such that maximizes the correlation.
}
