% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Voronoi.R
\name{distanceGrid}
\alias{distanceGrid}
\title{Grid Generation: Calculate tiles's distance to cnetroids in WVD}
\usage{
distanceGrid(centroids, grid_precluster, weight = F)
}
\arguments{
\item{centroids}{A data.frame, the coordinates (X, Y) of centroids in 2-D plate; should include "Weight" column as necessary input; can include "Class" as annotations}

\item{grid_precluster}{A data.frame, recording the tile coordinats (X, Y)}

\item{weight}{A logical value, whether doing weighted clustering or not}
}
\value{
A matrix including the distances of each tile to the centorids
}
\description{
Calculate the distance of each tile to different centroids. Can add weights on each centroids to apply the weighted clustering and thus generate the WVD
}
