% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMD.R
\name{EMDmat}
\alias{EMDmat}
\title{Earth Mover's Distance between clusters}
\usage{
EMDmat(coord, spatial.distribution, range.threshold = 5)
}
\arguments{
\item{coord}{Coordinates of tiles}

\item{spatial.distribution}{Cluster of tiles in a sparse form; should be a ntile-by-ncluster matrix}

\item{range.threshold}{Numeric, the maximal distance (in map units) over which locations are compared.}
}
\value{
A matrix with EMD between clusters; diagonals are 0
}
\description{
Calculate EMD between clusters; apply distance threshold to accelerate the computation
}
