% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kriging.R
\name{fitKriging}
\alias{fitKriging}
\title{Perform Kriging model on integrated ST data}
\usage{
fitKriging(
  spatial.data,
  gene.id,
  Kriging.cutoff = 600,
  Kriging.width = 30,
  plot.dir = "../plots/Kriging/",
  save.plot = T
)
}
\arguments{
\item{spatial.data}{Integrated ST data from integrateCoordinate.}

\item{gene.id}{Gene ID to perform Kriging. A character}

\item{Kriging.cutoff}{Kriging parameters setting the max range of model; a numeric value}

\item{Kriging.width}{Kriging parameters setting the width of bins when building the model; a numeric value}

\item{plot.dir}{Path to the plots outputs}

\item{save.plot}{Logical value, indicating whether do the plotting or not}
}
\value{
A list of Kriging models, containing the model parameters and model fits.
}
\description{
Second sub-function of SpatialKriging. Fit the Kriging model on the integrated ST data.
}
