% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Moran_I.R
\name{getMoran}
\alias{getMoran}
\title{Compute Moran's I on given gene}
\usage{
getMoran(gene, dist_inv)
}
\arguments{
\item{gene}{A vector of gene's expression value. Should correspond to the IDM}

\item{dist_inv}{The inverse distance matrix}
}
\value{
A 2-value vector containing the observed Moran's I and the p-value
}
\description{
On a given spatial coordinates, use the inverse distance matrix to calculate the Moran's I
}
