% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kriging.R
\name{integrateCoordinate}
\alias{integrateCoordinate}
\title{Integrate the ST data and spatial coordinates}
\usage{
integrateCoordinate(spatial.data, spatial.coord)
}
\arguments{
\item{spatial.data}{Spot-level spatial transcriptomics data; should be a spot-by-gene data frame}

\item{spatial.coord}{Spatial coordinates of the spots; should have same rownames with spatial.data and column names as 'imagerow' and 'imagecol'}
}
\value{
An object of class SpatialPointsDataFrame from sp::coordinates()
}
\description{
First sub-function of SpatialKriging. Combine the expression data and coordinates together for Kriging model functions.
}
