% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellMerge.R
\name{mergekNeighbors}
\alias{mergekNeighbors}
\title{Merge single cells into meta-cell based on their distance}
\usage{
mergekNeighbors(sc.data, k = 10)
}
\arguments{
\item{sc.data}{A data.frame as the scRNA-seq data}

\item{k}{The number of neighbors to choose. An integer}
}
\value{
A data.frame containing the new scRNA-seq data after merging
}
\description{
Find the neighbors of each single cells and merge them together into one meta-cell. Can be useful in blurring scRNA-seq data and obtaining the neighborhood information
}
