% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kriging.R
\name{predictKriging}
\alias{predictKriging}
\title{Kriging on Spatial Transcriptomics}
\usage{
predictKriging(
  spatial.data,
  new_spatial_grid,
  gene.id,
  Kriging.model,
  plot.dir = "../plots/Kriging/",
  save.plot = T
)
}
\arguments{
\item{spatial.data}{Integrated ST data from integrateCoordinate.}

\item{new_spatial_grid}{A data.frame containing the new grid coordinates. Generated from createGridDataframe}

\item{gene.id}{Gene ID to perform Kriging. A character}

\item{Kriging.model}{Variogram model fitted by gstat::fit.variogram. Can be obtained by fitKriging}

\item{plot.dir}{Path to the plots outputs}

\item{save.plot}{Logical value, indicating whether do the plotting or not}
}
\value{
An object from krige predict function. Contain the prediction, variance and coordinates on new gird.
}
\description{
Last sub-function of SpatialKriging. Predict the spatial profile of given gene on the newly-created grid.
}
