% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMD.R
\name{simulateST}
\alias{simulateST}
\title{Simulate spatial transcriptomics data by markov random field}
\usage{
simulateST(
  height = 100,
  width = 100,
  ncluster = 4,
  max.norm = 3,
  param.mat = c(0, -0.1, -0.2, -0.4, -0.1, 0, -0.15, -0.35, -0.2, -0.15, 0, -0.1, -0.4,
    -0.35, -0.1, 0) * 0.7,
  ncycle = 20,
  plot = T,
  random.seed = 1111
)
}
\arguments{
\item{height}{Number of tiles for height}

\item{width}{Number of tiles for width}

\item{ncluster}{Number of cluster in simulation}

\item{max.norm}{A numeric value. All points with norm ≤ max.norm are included.}

\item{param.mat}{Parameter matrix for simulating spatial relationships between clusters; should be ncluster-by-ncluster}

\item{ncycle}{Cycles to run in MRF simulation}

\item{plot}{A logical value, whether draw the plot}

\item{random.seed}{Random seed to use}
}
\value{
A matrix with the sampled field.
}
\description{
With given tuned parameters, this function performs a simulation of ST data that includes spatial
relationships between clusters. It could be used as the simulation data for any ST-related algorithms.
}
