% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kriging.R
\name{SpatialKriging}
\alias{SpatialKriging}
\title{Kriging on Spatial Transcriptomics}
\usage{
SpatialKriging(
  spatial.data,
  spatial.coord,
  gene.id,
  predict.grid.size = 20,
  Kriging.cutoff = 600,
  Kriging.width = 30,
  plot.dir = "../plots/Kriging/",
  save.plot = T
)
}
\arguments{
\item{spatial.data}{Spot-level spatial transcriptomics data; should be a spot-by-gene data frame}

\item{spatial.coord}{Spatial coordinates of the spots; should have same rownames with spatial.data}

\item{gene.id}{Gene ID to perform Kriging. A character}

\item{predict.grid.size}{The size of the finer grid; a numeric value used in createGridDataframe}

\item{Kriging.cutoff}{Kriging parameters setting the max range of model; a numeric value}

\item{Kriging.width}{Kriging parameters setting the width of bins when building the model; a numeric value}

\item{plot.dir}{Path to the plots outputs}

\item{save.plot}{Logical value, indicating whether do the plotting or not}
}
\value{
An object from krige predict function. Contain the prediction, variance and coordinates on new gird.
}
\description{
Based on the spot-level spatial transcriptomics profiles, apply the ordinary Kriging on given genes. The function includes 4 sub-functions that perform different tasks: integrateCoordinate, fitKriging, createGridDataframe and predictKriging.
}
