% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kriging.R
\name{SpatialKrigingCV}
\alias{SpatialKrigingCV}
\title{Cross-Validation SpatialKriging on spot-level ST data}
\usage{
SpatialKrigingCV(
  spatial.data,
  spatial.coord,
  gene.id,
  fraction = 0.1,
  Kriging.cutoff = 600,
  Kriging.width = 30,
  random.seed = NULL
)
}
\arguments{
\item{spatial.data}{Spot-level spatial transcriptomics data; should be a spot-by-gene data frame}

\item{spatial.coord}{Spatial coordinates of the spots; should have same rownames with spatial.data}

\item{gene.id}{Gene ID to perform Kriging. A character}

\item{fraction}{The faction of test spots. A numeric value from 0 to 1}

\item{Kriging.cutoff}{Kriging parameters setting the max range of model; a numeric value}

\item{Kriging.width}{Kriging parameters setting the width of bins when building the model; a numeric value}

\item{random.seed}{The random seed used in sampling training spots}
}
\value{
The correlation between prediction and original values. A numeric score from -1 to 1
}
\description{
Evaluate the performance of Kriging model by doing CV on spot-level data. Fit the model on a fraction of spots and predict the rest on the expression profile of given gene. The result is measured by the correlation between the origianl and predicted expresion across the space.
}
