% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spot_coord}
\alias{spot_coord}
\title{spot_coord}
\format{
An object of class \code{data.frame} with 722 rows and 2 columns.
}
\usage{
spot_coord
}
\description{
This is data to be included in my package
}
\author{
Tianxiao Zhao \email{Tianxiao.Zhao@nyulangone.org}
}
\keyword{coordinates}
\keyword{of}
\keyword{spots}
