% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ST_expr_data}
\alias{ST_expr_data}
\title{ST_expr_data}
\format{
An object of class \code{data.frame} with 722 rows and 247 columns.
}
\usage{
ST_expr_data
}
\description{
The Spatial Transcriptomics data, which is a spot-by-gene data frame
}
\author{
Tianxiao Zhao \email{Tianxiao.Zhao@nyulangone.org}
}
\keyword{Spatial}
\keyword{Transcriptomics}
\keyword{data}
