% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Voronoi_centroid}
\alias{Voronoi_centroid}
\title{Voronoi_centroid}
\format{
An object of class \code{data.frame} with 13 rows and 4 columns.
}
\usage{
Voronoi_centroid
}
\description{
A data frame including the coorinates of cell state centroids after MDS. Also includes the weights to generate the WVD
}
\author{
Tianxiao Zhao \email{Tianxiao.Zhao@nyulangone.org}
}
\keyword{Voronoi}
\keyword{centroids}
