% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Voronoi_connections}
\alias{Voronoi_connections}
\title{Voronoi_connections}
\format{
An object of class \code{data.frame} with 9 rows and 5 columns.
}
\usage{
Voronoi_connections
}
\description{
A data frame including the connections of cell state centroids after MDS. Also includes the weights to generate the WVD
}
\author{
Tianxiao Zhao \email{Tianxiao.Zhao@nyulangone.org}
}
\keyword{Voronoi}
\keyword{connections}
