% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Voronoi.R
\name{WVDpreprocess}
\alias{WVDpreprocess}
\title{Pre-process the data for Weighted Voronoi Diagram (WVD)}
\usage{
WVDpreprocess(
  mds.mat,
  spatial.distribution,
  centroid.weight = 0.5,
  connection.weight = 0.3,
  grid.bound.factor = 1.1,
  max.range = 10,
  ngrid = 100
)
}
\arguments{
\item{mds.mat}{Matrix containing the position of clusters after MDS}

\item{spatial.distribution}{Cluster of tiles in a sparse form; should be a ntile-by-ncluster matrix}

\item{centroid.weight}{Numeric, weights in WVD when creating the Voronoi cells}

\item{connection.weight}{Numeric, connections whose weights greater than given values will be ploted}

\item{grid.bound.factor}{Numeric bigger than 1, how large lattice will be compared to the max value in MDS}

\item{max.range}{The max range when clustering; grids beyond this value will be clustered as "Outlier"}

\item{ngrid}{Number of grids in lattice}
}
\value{
A list, including the information of 1) centroids; 2) connections; 3) Voronoi cells.
}
\description{
From the multidimensional scaling matrix, generate the centroids, connections and Voronoi cells
}
