% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellAlignment.R
\name{CalcCor}
\alias{CalcCor}
\title{Calculate correlations between scRNA-seq data and spatial transcriptomics}
\usage{
CalcCor(sc.data, ST.data.column, method.use = "pearson")
}
\arguments{
\item{sc.data}{A data.frame indicating the scRNA-seq data; each column is a cell and each row is a gene}

\item{ST.data.column}{A vector containing the gene expression of a given spot}

\item{method.use}{a character string indicating which correlation coefficient (or covariance) is to be computed: "pearson", "kendall" or "spearman"}
}
\value{
A vector of correlations between cells and one spot
}
\description{
A brief function to calculate the correlation between all cells compared to one spot in ST data
}
