% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Voronoi.R
\name{clusterGrid}
\alias{clusterGrid}
\title{Grid Generation: Cluster tiles based on the distance}
\usage{
clusterGrid(distance.grid)
}
\arguments{
\item{distance.grid}{A matrix from SpatialAbstractMap::distanceGrid; include the distances of each tile to the centorids}
}
\value{
A data.frame including the cluster (centroid) and the corresponding distance (weighted or not)
}
\description{
Find the nearest centroids for every tile as the cluster results.
}
